pyproject.toml
[project]
name = "myfastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "google-genai>=1.56.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
